Moved from Vector3f to Vector3d for precision + total calculations

+++ simulator/src/com/verictas/pos/simulator/Main.java
@@ -5,10 +5,10 @@ import javax.vecmath.*;
 public class Main {
 
     public static void main(String[] args) {
-        Object object1 = new Object(1000, new Vector3f(1,2,3), new Vector3f(0,4,3));
-        Object object2 = new Object(200, new Vector3f(2,38,2), new Vector3f(3,4,5));
-        Object object3 = new Object(200, new Vector3f(2,-20,2), new Vector3f(3,4,5));
-        Object object4 = new Object(200, new Vector3f(2,4,2), new Vector3f(3,4,5));
+        Object object1 = new Object(1000, new Vector3d(1,2,3), new Vector3d(0,4,3));
+        Object object2 = new Object(200, new Vector3d(2,38,2), new Vector3d(3,4,5));
+        Object object3 = new Object(200, new Vector3d(2,-20,2), new Vector3d(3,4,5));
+        Object object4 = new Object(200, new Vector3d(2,4,2), new Vector3d(3,4,5));
 
         // Make a list of all the objects
         Object[] objects = {object1, object2, object3, object4};
